+++ MenuProject/src/com/s7design/menu/dataclasses/Rate.java
@@ -2,7 +2,7 @@ package com.s7design.menu.dataclasses;
 
 public class Rate {
 
-	public double tax;
-	public double mintip;
-	public double maxtip;
+	public float tax;
+	public float mintip;
+	public float maxtip;
 }
