+++ MenuProject/src/com/s7design/menu/volley/responses/GetDiscountResponse.java
@@ -2,6 +2,6 @@ package com.s7design.menu.volley.responses;
 
 public class GetDiscountResponse extends GsonResponse {
 
-	public double discount;
+	public float discount;
 
 }
