+++ MenuProject/src/com/s7design/menu/volley/responses/GetReceiptInfoResponse.java
@@ -5,10 +5,10 @@ import com.s7design.menu.dataclasses.Item;
 public class GetReceiptInfoResponse extends GsonResponse {
 
 	public String response;
-	public double orderprice;
-	public double discount;
-	public double tip;
-	public double tax;
+	public float orderprice;
+	public float discount;
+	public float tip;
+	public float tax;
 	public String date;
 	public String time;
 	public Item[] items;
